Move mapEventDtoToEvent method into EventDTO
